Rename frontend to next-api, update refs
